Refactor Complete
