+++ app/frontend/src/components/QuestionInput/QuestionInput.tsx
@@ -3,13 +3,13 @@ import { useMsal } from "@azure/msal-react";
 import { IStyleFunctionOrObject, Stack, TextField } from "@fluentui/react";
 import { Button, Tooltip, Spinner, ToggleButton } from "@fluentui/react-components";
 import { Toggle } from "@fluentui/react/lib/Toggle";
-import { Send24Filled, Dismiss24Filled, Filter24Filled, FilterDismiss24Filled } from "@fluentui/react-icons";
 import { isLoggedIn, requireAccessControl } from "../../authConfig";
 
 import styles from "./QuestionInput.module.css";
 import UploadFiles from "../UploadFiles/UploadFiles";
 import { useMutation } from "react-query";
 import { IRemoveRequest, IRemoveResponse, removeFilesApi } from "../../api";
+import { Send24Filled } from "@fluentui/react-icons";
 
 interface Props {
     onSend: (question: string) => void;
@@ -22,31 +22,16 @@ interface Props {
 
 export const QuestionInput = ({ onSend, setDocFilter, disabled, placeholder, clearOnSend, initQuestion }: Props) => {
     const [question, setQuestion] = useState<string>("");
+    const [filter, setFilter] = useState<string | undefined>();
 
-    const [isOn, setIsOn] = useState<boolean>(true);
-    const [uploadedFiles, setUploadedFiles] = useState<File[] | null>(null);
-
-    const mutation = useMutation<any, Error, { files: string[]; idToken?: string }>(({ files }) => removeFilesApi(files, undefined), {
-        onSettled: (data, error) => {
-            if (error) {
-                console.log("Deletion failed:", error);
-            } else {
-                console.log("Deletion Successful");
-            }
-        }
-    });
+    const setDocumentFilter = (filter: string | undefined) => {
+        setDocFilter(filter);
+        setFilter(filter);
+    };
 
     useEffect(() => {
         initQuestion && setQuestion(initQuestion);
     }, [initQuestion]);
-
-    useEffect(() => {
-        if (isOn && uploadedFiles) {
-            constructAndSetDocFilter();
-        } else {
-            setDocFilter(undefined);
-        }
-    }, [isOn, uploadedFiles]);
 
     const sendQuestion = () => {
         if (disabled || !question.trim()) {
@@ -75,35 +60,6 @@ export const QuestionInput = ({ onSend, setDocFilter, disabled, placeholder, cle
         }
     };
 
-    const constructAndSetDocFilter = () => {
-        if (uploadedFiles && isOn) {
-            const names = uploadedFiles.map(file => file.name);
-
-            console.log("Names", names);
-            console.log("Uploaded files", uploadedFiles);
-
-            if (names.length === 1) {
-                setDocFilter(names[0]);
-            } else {
-                const str: string = names.join(" OR ");
-                setDocFilter(str);
-            }
-        }
-    };
-
-    const removeFiles = async () => {
-        const response: IRemoveResponse = await mutation.mutateAsync({
-            files: uploadedFiles?.map(file => file.name) as string[]
-        });
-        setUploadedFiles(null);
-        setDocFilter(undefined);
-    };
-
-    const removeFilter = () => {
-        setDocFilter(undefined);
-        setIsOn(!isOn);
-    };
-
     const { instance } = useMsal();
     const disableRequiredAccessControl = requireAccessControl && !isLoggedIn(instance);
     const sendQuestionDisabled = disabled || !question.trim() || disableRequiredAccessControl;
@@ -114,7 +70,7 @@ export const QuestionInput = ({ onSend, setDocFilter, disabled, placeholder, cle
 
     return (
         <Stack className={styles.inputContainer}>
-            {isOn && uploadedFiles && <p>Only retrieving information from: {uploadedFiles[0].name}</p>}
+            {filter && <p>Only retrieving information from: {filter}</p>}
             <Stack horizontal className={styles.inputUpload}>
                 <Stack horizontal className={styles.questionInputContainer}>
                     <TextField
@@ -139,26 +95,7 @@ export const QuestionInput = ({ onSend, setDocFilter, disabled, placeholder, cle
                         </Tooltip>
                     </div>
                 </Stack>
-                <div className={styles.uploadButtonContainer}>
-                    {/* <UploadFiles setUploadedFiles={setUploadedFiles} /> */}
-                    {!uploadedFiles && <UploadFiles setUploadedFiles={setUploadedFiles} />}
-                    {uploadedFiles && (
-                        <div>
-                            {isOn ? (
-                                <Tooltip content="Disable filter" relationship="label">
-                                    <Button size="large" icon={<FilterDismiss24Filled primaryFill="rgba(115, 118, 225, 1)" />} onClick={removeFilter} />
-                                </Tooltip>
-                            ) : (
-                                <Tooltip content="Enable filter" relationship="label">
-                                    <Button size="large" icon={<Filter24Filled primaryFill="rgba(115, 118, 225, 1)" />} onClick={removeFilter} />
-                                </Tooltip>
-                            )}
-                            <Tooltip content="Remove Files" relationship="label">
-                                <Button size="large" icon={<Dismiss24Filled primaryFill="rgba(115, 118, 225, 1)" />} onClick={removeFiles} />
-                            </Tooltip>
-                        </div>
-                    )}
-                </div>
+                <UploadFiles setDocFilter={setDocumentFilter} />
             </Stack>
         </Stack>
     );
