+++ app/frontend/src/data/Experiment_Victoria/evaluate_parameters 1.json
@@ -1,6 +1,6 @@
 {
     "evaluation_gpt_model": "gpt-35-turbo",
-    "evaluation_timestamp": 1708706573,
+    "evaluation_timestamp": 1709029531,
     "testdata_path": "C:\\Users\\ZU583EQ\\Projects\\ai-rag-chat-evaluator\\example_input\\qa.jsonl",
     "target_url": "http://127.0.0.1:50505/chat",
     "target_parameters": {
