--- app/frontend/src/data/Experiment_Victoria/summary.json
@@ -1,27 +1,0 @@
-{
-    "relevance": {
-        "mean_rating": 3.28,
-        "pass_count": 32,
-        "pass_rate": 0.53
-    },
-    "coherence": {
-        "mean_rating": 3.98,
-        "pass_count": 55,
-        "pass_rate": 0.92
-    },
-    "groundedness": {
-        "mean_rating": 4.53,
-        "pass_count": 53,
-        "pass_rate": 0.88
-    },
-    "latency": {
-        "mean": 8.39,
-        "max": 13.175334,
-        "min": 5.822337
-    },
-    "answer_length": {
-        "mean": 655.82,
-        "max": 2321,
-        "min": 67
-    }
-}
